oauth works and redirects to user page, storing user context
