+++ src/display/components/UserPage.jsx
@@ -2,8 +2,14 @@ import React, { useEffect, useState, useContext } from 'react';
 import { UserContext } from './context/UserContext.jsx';
 
 function UserPage (props) {
+  const { user } = useContext(UserContext);
+
   return (
-    <div>USER PAGE HOLD OVER</div>
+    <div>
+      {user.name}
+      {user.u_id}
+      {user.email}
+    </div>
   )
 }
